Move: @component/new/... 디렉토리에서 new 디렉토리 삭제 후 컴포넌트 전부 이동
