+++ client/src/component/list/PhotoList.tsx
@@ -4,8 +4,8 @@ import { useQueryClient } from '@tanstack/react-query';
 import * as queryKey from '@api/queryKey';
 import usePhotosQuery, { FilterType } from '@api/query/photo/usePhotosQuery';
 import NextPageFetcher from '@component/list/content/NextPageFetcher';
-import PhotocardItem from '@component/photocard/new/item/PhotocardItem';
-import SkeletonPhotoCardItem from '@component/photocard/new/item/SkeletonPhotocardItem';
+import PhotocardItem from '@component/photocard/item/PhotocardItem';
+import SkeletonPhotoCardItem from '@component/photocard/item/SkeletonPhotocardItem';
 import { IconType } from '@type/icon';
 import { State as FilterState } from '@component/search/content/filter/reducer';
 import { State as KeywordState } from '@component/search/content/keyword/reducer';
