+++ client/src/component/photocard/info/PhotoInfo.tsx
