+++ client/src/component/photocard/info/SkeletonPhotoInfo.tsx
