+++ client/src/component/photocard/info/_styles.ts
