+++ client/src/component/photocard/item/SkeletonPhotocardItem.tsx
