+++ client/src/component/photocard/item/_styles.ts
