+++ client/src/component/shipping/request/RequestVoucherInfo.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Card, { StylesProps as CardStyles } from '@component/card/basic/Card';
 import CardHeader from '@component/card/basic/CardHeader';
 import CardBody from '@component/card/basic/CardBody';
-import PhotocardItem from '@component/photocard/new/item/PhotocardItem';
+import PhotocardItem from '@component/photocard/item/PhotocardItem';
 import { ItemSection } from '@component/list/content/_styles';
 import { VoucherItem } from '@type/voucher';
 import { IconType } from '@type/icon';
