+++ client/src/page/admin/photo/detail/content/PhotoInfo.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { faPenToSquare } from '@fortawesome/free-solid-svg-icons';
 import { Photo } from '@type/photo';
 import Button from '@component/form/Button';
-import PhotoInfoComponent from '@component/photocard/new/info/PhotoInfo';
+import PhotoInfoComponent from '@component/photocard/info/PhotoInfo';
 
 interface Props {
   photo: Photo;
