+++ client/src/page/admin/voucher/writer/content/photo/PhotoList.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useQueries } from '@tanstack/react-query';
 import * as queryKey from '@api/queryKey';
 import { fetchPhotoDetail } from '@api/api/photo';
-import SkeletonPhotoCardItem from '@component/photocard/new/item/SkeletonPhotocardItem';
+import SkeletonPhotoCardItem from '@component/photocard/item/SkeletonPhotocardItem';
 import { State, Action } from '../../reducer';
 import { ItemSection } from '@component/list/content/_styles';
 import { PhotoItemType } from './PhotoItem';
