+++ client/src/page/user/trade/writer/content/voucher/CardSection.tsx
@@ -3,8 +3,8 @@ import useVoucherQuery from '@api/query/voucher/useVoucherQuery';
 import * as queryKey from '@api/queryKey';
 import { faAdd } from '@fortawesome/free-solid-svg-icons';
 import { ModalHookType } from '@hook/useModal';
-import PhotoInfo from '@component/photocard/new/info/PhotoInfo';
-import SkeletonPhotoInfo from '@component/photocard/new/info/SkeletonPhotoInfo';
+import PhotoInfo from '@component/photocard/info/PhotoInfo';
+import SkeletonPhotoInfo from '@component/photocard/info/SkeletonPhotoInfo';
 import Card from '@component/card/basic/Card';
 import CardHeader from '@component/card/basic/CardHeader';
 import CardBody from '@component/card/basic/CardBody';
