+++ client/src/page/user/voucher/detail/Success.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import VoucherInfo from '@component/voucher/info/VoucherInfo';
-import PhotoInfo from '@component/photocard/new/info/PhotoInfo';
+import PhotoInfo from '@component/photocard/info/PhotoInfo';
 import { ResType } from '@api/query/voucher/useVoucherQuery';
 import ButtonSection from './content/ButtonSection';
